+++ packages/adminui-pub/CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+ - `UBCommand/getNavshortcutCommandText` param shortcutCode in returned commandConfig 
 
 ### Changed